Mon, Jul 22, 2024, 4:35 PM -04:00

+++ week4_pset/filter-more/helpers.c
@@ -11,7 +11,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
         for (int j = 0; j < width; j++)
         {
             // Defining a variable to hold the average of RGB values for each pixel
-            int averageRGB = 0;
+            float averageRGB = 0.0;
             averageRGB += image[i][j].rgbtRed;
             averageRGB += image[i][j].rgbtGreen;
             averageRGB += image[i][j].rgbtBlue;
